Migrate component me-gallery-view
